Move new reducer out of content
