+++ src/app/types.ts
@@ -22,6 +22,7 @@ import PromiseCollector from '../helpers/PromiseCollector';
 import { State as authState } from './auth/types';
 import { State as contentState } from './content/types';
 import { State as errorsState } from './errors/types';
+import { State as featureFlagsState } from './featureFlags/types';
 import { State as headState } from './head/types';
 import { State as navigationState } from './navigation/types';
 import { State as notificationState } from './notifications/types';
@@ -29,6 +30,7 @@ import { State as notificationState } from './notifications/types';
 export interface AppState {
   content: contentState;
   errors: errorsState;
+  experiments: featureFlagsState;
   head: headState;
   auth: authState;
   navigation: navigationState;
